+++ src/github/client/org.rs
@@ -7,7 +7,7 @@ impl GithubClient {
 			"{}/orgs/{}/members/{}",
 			self.github_api_url, org, username
 		);
-		let status = self.client.get_status(url).await?;
+		let status = self.get_status(url).await?;
 		// https://docs.github.com/en/rest/orgs/members#check-organization-membership-for-a-user--code-samples
 		Ok(status == 204)
 	}
